Refractor files paths
