+++ src/components/events/ImageHeaderComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { primary } from '../utils/Colors';
+import { primary } from '../../utils/Colors';
 import PropTypes from 'prop-types';
 
 import { withStyles } from '@material-ui/core/styles';
@@ -25,10 +25,10 @@ const styles = theme => ({
   },
 });
 
-class EventHeaderComponent extends Component {
+class ImageHeaderComponent extends Component {
   
   render() {
-    const { siteTitle, siteSecondaryTitle } = this.props;
+    const { siteTitle, location, siteDescription, date } = this.props;
     const { classes } = this.props;
 
     return (
@@ -38,8 +38,12 @@ class EventHeaderComponent extends Component {
             {siteTitle}
           </Typography>
 
+          <Typography variant="h6" className={classes.whiteText}>
+            {location} ∙ {date}
+          </Typography>
+
           <Typography variant="body1" className={classes.whiteText}>
-            {siteSecondaryTitle}
+            {siteDescription}
           </Typography>
         </div>
       </div>
@@ -47,9 +51,11 @@ class EventHeaderComponent extends Component {
   }
 }
 
-EventHeaderComponent.propTypes = {
+ImageHeaderComponent.propTypes = {
   siteTitle: PropTypes.string.isRequired,
-  siteSecondaryTitle: PropTypes.string.isRequired,
+  location: PropTypes.string.isRequired,
+  date: PropTypes.string.isRequired,
+  siteDescription: PropTypes.string.isRequired,
 };
 
-export default withStyles(styles)(EventHeaderComponent);
+export default withStyles(styles)(ImageHeaderComponent);
