+++ src/components/home/featured_events/FeaturedEventCarouselComponent.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import BannerImage from './image'
+import BannerImage from '../../BannerImage'
 
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
