+++ src/components/home/looking_ahead/ValuesComponent.js
